refactor: share test utils in freenet_core
