+++ crates/core/src/lib.rs
@@ -71,5 +71,4 @@ pub mod dev_tool {
     pub use wasm_runtime::{ContractStore, DelegateStore, Runtime, SecretsStore, StateStore};
 }
 
-#[cfg(test)]
 pub mod test_utils;
